[kernel/libc] Small source code cleanup

+++ elkscmd/basic/basic.c
@@ -860,7 +860,7 @@ int nextToken()
                 gotDecimal = true;
             else {
                 *tokenOut++ = TOKEN_INTEGER;
-                *(long*)tokenOut = (long)val;
+                *(long*)tokenOut = val;
                 tokenOut += sizeof(long);
             }
         }
